ai_video_creator/requirements.txt
# AI服务
openai>=1.3.0  # GPT文本生成
stability-sdk>=0.8.0  # Stable Diffusion图像生成
baidu-aip>=4.16.3  # 百度AI平台SDK

# 视频处理
moviepy==1.0.3  # 视频编辑
decorator>=4.0.11  # moviepy依赖
imageio-ffmpeg>=0.4.3  # moviepy依赖
imageio>=2.5.0  # moviepy依赖
Pillow>=10.0.0  # 图像处理
numpy>=1.24.0  # 数值计算

# 工具
python-dotenv>=0.19.0  # 环境变量
requests>=2.26.0  # HTTP请求
urllib3<2.0.0  # 解决SSL兼容性问题
tqdm>=4.65.0  # 进度条
chardet>=5.0.0  # 字符编码检测
